Village Manager

This object manages all of the villages. It acts as an interface to them

Methods

initialize()

Creates a new manager

get_villages()

Gets all of the villages that the manager has

get_village()

Retrieves a specific village from the manager, by name

add_village()

Adds a village to the manager

Public fields

villages

A list of village objects Creates a new, village manager

Methods

Public methods

village_manager$new()

village_manager$get_villages()

village_manager$get_village()

village_manager$add_resource()

village_manager$clone()

Method new()

Get a new instance of a village_manager Gets all of the managed villages

Usage

village_manager$new(villages)

Method get_villages()

Usage

village_manager$get_villages()

Returns

A list of resources Gets a village given a village name

Method get_village()

Usage

village_manager$get_village(name)

Arguments

name

The name of the requested village

Returns

A village object Adds a village to the manager.

Method add_resource()

Usage

village_manager$add_resource(...)

Arguments

...

The villages to add

Returns

None

Method clone()

The objects of this class are cloneable with this method.

Usage

village_manager$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.